playlist preservation

+++ Pandorian/frmMain.vb
@@ -93,9 +93,11 @@ Public Class frmMain
     Public Sub ClearSession()
         If Not IsNothing(Pandora) Then
             Pandora.ClearSession(Settings.pandoraOne)
-            'File.Delete(TempPath + APIFile)
             SavePandoraObject()
-            CleanUp()
+
+            Dim bgwCleanCache As New BackgroundWorker
+            AddHandler bgwCleanCache.DoWork, AddressOf CleanUpCache
+            bgwCleanCache.RunWorkerAsync()
         End If
     End Sub
 
@@ -149,7 +151,6 @@ Public Class frmMain
                 End If
             Case 10
                 volSlider.Visible = True
-                Application.DoEvents()
                 VolLastChangedOn = Now
                 Try
                     volSlider.Value = volSlider.Value - 10
@@ -158,7 +159,6 @@ Public Class frmMain
                 End Try
             Case 11
                 volSlider.Visible = True
-                Application.DoEvents()
                 VolLastChangedOn = Now
                 Try
                     volSlider.Value = volSlider.Value + 10
@@ -254,8 +254,10 @@ Public Class frmMain
         Pandora.ClearSession(Settings.pandoraOne)
         Return False
     End Function
+
     Sub LoadStationList()
         If Not Pandora.AvailableStations.Count = 0 Then
+
             Dim FoundLastPlayedStation As Boolean
             Dim Stations As New SortedDictionary(Of String, String)
             For Each Station In Pandora.AvailableStations
@@ -369,9 +371,8 @@ Public Class frmMain
 
         If Not IsNothing(Pandora.CurrentStation.CurrentSong.PreviousSong) Then
             Spinner.Visible = True
-            prgBar.Value = 0
-            prgBar.Update()
             Application.DoEvents()
+            prgBar.Value = 0
             ResumePlaying = True
             Pandora.CurrentStation.CurrentSong = Pandora.CurrentStation.CurrentSong.PreviousSong
             Execute(Sub() PlayCurrentSong(), "PlayPreviousSong") 'don't ever change this string [handling-of-expired-url]
@@ -386,9 +387,8 @@ Public Class frmMain
     Sub PlayNextSong()
 
         Spinner.Visible = True
-        prgBar.Value = 0
-        prgBar.Update()
         Application.DoEvents()
+        prgBar.Value = 0
         ResumePlaying = True
 
         Try
@@ -411,13 +411,13 @@ Public Class frmMain
                         Bass.BASS_ChannelSetPosition(Stream, 0)
                         Bass.BASS_ChannelPlay(Stream, False)
                         Spinner.Visible = False
+                        Application.DoEvents()
                         tbLog.AppendText("Global skip limit reached. Replaying current song..." + vbCrLf)
                         Pandora.SkipLimitReached = True
                         Pandora.SkipLimitReachedAt = Now
                         ddStations.Enabled = False
                         btnNext.Enabled = False
                         btnNext.BackColor = Color.DarkGray
-                        Application.DoEvents()
                         Exit Sub
                     End If
                     Throw x
@@ -428,10 +428,10 @@ Public Class frmMain
                 Bass.BASS_ChannelSetPosition(Stream, 0)
                 Bass.BASS_ChannelPlay(Stream, False)
                 Spinner.Visible = False
+                Application.DoEvents()
                 ddStations.Enabled = False
                 btnNext.Enabled = False
                 btnNext.BackColor = Color.DarkGray
-                Application.DoEvents()
                 Exit Sub
             End If
         End Try
@@ -539,7 +539,6 @@ Public Class frmMain
 
             BPMCounter.Reset(44100)
 
-            Application.DoEvents()
             song.PlayingStartTime = Now
             song.AudioDurationSecs = SongDurationSecs()
             ShareTheLove()
@@ -741,20 +740,44 @@ Public Class frmMain
         Return Text
     End Function
 
-    Private Sub CleanUp()
-        For Each f In Directory.GetFiles(Path.GetTempPath, "*.cover")
-            Try
-                File.Delete(f)
-            Catch ex As Exception
-                'do null
-            End Try
+    Private Sub CleanUpCache(sender As Object, e As DoWorkEventArgs)
+        Dim tmp = Path.GetTempPath
+        Dim keep As New List(Of String)
+        For Each stn In Pandora.AvailableStations
+            For Each s In stn.PlayList.ToArray
+                keep.Add(tmp + s.Token + ".stream")
+            Next
         Next
-        For Each f In Directory.GetFiles(Path.GetTempPath, "*.stream")
-            Try
-                File.Delete(f)
-            Catch ex As Exception
-                'do null
-            End Try
+
+        Dim keepFiles = keep.ToArray
+        Dim files = Directory.GetFiles(tmp, "*.stream")
+
+        For Each f In files
+            If Array.IndexOf(keepFiles, f) < 0 Then
+                Try
+                    File.Delete(f)
+                Catch ex As Exception
+                    'do null
+                End Try
+            End If
+        Next
+
+        Dim keepCovers As New List(Of String)
+        For Each f In keepFiles
+            keepCovers.Add(f.Replace(".stream", ".cover"))
+        Next
+
+        files = Directory.GetFiles(tmp, "*.cover")
+        keepFiles = keepCovers.ToArray
+
+        For Each f In files
+            If Array.IndexOf(keepFiles, f) < 0 Then
+                Try
+                    File.Delete(f)
+                Catch ex As Exception
+                    'do null
+                End Try
+            End If
         Next
     End Sub
 
@@ -780,7 +803,6 @@ Public Class frmMain
         populateSleepTimes()
         AddHandler SystemEvents.PowerModeChanged, AddressOf PowerModeChanged
         AddHandler SystemEvents.SessionEnding, AddressOf MachineShutDown
-        Application.DoEvents()
     End Sub
 
     Private Sub TrayIcon_MouseClick(sender As Object, e As MouseEventArgs) Handles TrayIcon.MouseClick
@@ -810,7 +832,6 @@ Public Class frmMain
     End Sub
 
     Private Sub frmMain_Shown(sender As Object, e As EventArgs) Handles Me.Shown
-        Application.DoEvents()
         Execute(Sub() RunNow(), "frmMain_Shown.RunNow")
     End Sub
 
@@ -1102,9 +1123,7 @@ Public Class frmMain
     Private Sub ReLoginToPandora()
         Spinner.Visible = True
         Application.DoEvents()
-        Pandora.ClearSession(Settings.pandoraOne)
-        CleanUp()
-        SavePandoraObject()
+        ClearSession()
         Execute(Sub() RunNow(), "ReLoginToPandora.RunNow")
     End Sub
 
@@ -1412,7 +1431,6 @@ Public Class frmMain
 
             HideSongInfo = True
             SongInfo.Hide()
-            Application.DoEvents()
 
             If Not IsNothing(Pandora.CurrentStation.CurrentSong) And Not IsNothing(Pandora.CurrentStation) Then
                 tmiStationTitle.Text = Pandora.CurrentStation.Name.Replace("&", "&&")
